GithubIssueRepositoryのFakeがないため、既存のUIテストが失敗するので修正
